notas,toastr,etc

+++ database/migrations/2018_07_26_232343_create_tb_client_record_items_table.php
@@ -20,7 +20,7 @@ class CreateTbClientRecordItemsTable extends Migration
             $table->integer('client_record_id')->unsigned();
             $table->foreign('client_record_id')->references('id')->on('tb_client_records')->onDelete('cascade');
             $table->integer('quantity');
-            $table->double('item_total', 10, 2);
+            $table->decimal('item_total');
             $table->timestamps();
         });
     }
